+++ pkg/server/handlers/routes_test.go
@@ -28,6 +28,7 @@ import (
 	"github.com/dnote/dnote/pkg/assert"
 	"github.com/dnote/dnote/pkg/clock"
 	"github.com/dnote/dnote/pkg/server/database"
+	"github.com/dnote/dnote/pkg/server/mailer"
 	"github.com/dnote/dnote/pkg/server/testutils"
 	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
@@ -182,7 +183,6 @@ func TestGetCredential(t *testing.T) {
 }
 
 func TestAuthMiddleware(t *testing.T) {
-
 	defer testutils.ClearData()
 
 	user := testutils.SetupUserData()
@@ -296,7 +296,6 @@ func TestAuthMiddleware(t *testing.T) {
 }
 
 func TestAuthMiddleware_ProOnly(t *testing.T) {
-
 	defer testutils.ClearData()
 
 	user := testutils.SetupUserData()
@@ -385,7 +384,6 @@ func TestAuthMiddleware_ProOnly(t *testing.T) {
 }
 
 func TestTokenAuthMiddleWare(t *testing.T) {
-
 	defer testutils.ClearData()
 
 	user := testutils.SetupUserData()
@@ -515,7 +513,6 @@ func TestTokenAuthMiddleWare(t *testing.T) {
 }
 
 func TestTokenAuthMiddleWare_ProOnly(t *testing.T) {
-
 	defer testutils.ClearData()
 
 	user := testutils.SetupUserData()
@@ -688,6 +685,88 @@ func TestNotSupportedVersions(t *testing.T) {
 
 			// test
 			assert.Equal(t, res.StatusCode, http.StatusGone, "status code mismatch")
+		})
+	}
+}
+
+func TestNewRouter_AppValidate(t *testing.T) {
+	testCases := []struct {
+		app         App
+		expectedErr error
+	}{
+		{
+			app: App{
+				DB:               &gorm.DB{},
+				Clock:            clock.NewMock(),
+				StripeAPIBackend: nil,
+				EmailTemplates:   mailer.Templates{},
+				EmailBackend:     &testutils.MockEmailbackendImplementation{},
+				WebURL:           "http://mock.url",
+			},
+			expectedErr: nil,
+		},
+		{
+			app: App{
+				DB:               nil,
+				Clock:            clock.NewMock(),
+				StripeAPIBackend: nil,
+				EmailTemplates:   mailer.Templates{},
+				EmailBackend:     &testutils.MockEmailbackendImplementation{},
+				WebURL:           "http://mock.url",
+			},
+			expectedErr: ErrEmptyDB,
+		},
+		{
+			app: App{
+				DB:               &gorm.DB{},
+				Clock:            nil,
+				StripeAPIBackend: nil,
+				EmailTemplates:   mailer.Templates{},
+				EmailBackend:     &testutils.MockEmailbackendImplementation{},
+				WebURL:           "http://mock.url",
+			},
+			expectedErr: ErrEmptyClock,
+		},
+		{
+			app: App{
+				DB:               &gorm.DB{},
+				Clock:            clock.NewMock(),
+				StripeAPIBackend: nil,
+				EmailTemplates:   nil,
+				EmailBackend:     &testutils.MockEmailbackendImplementation{},
+				WebURL:           "http://mock.url",
+			},
+			expectedErr: ErrEmptyEmailTemplates,
+		},
+		{
+			app: App{
+				DB:               &gorm.DB{},
+				Clock:            clock.NewMock(),
+				StripeAPIBackend: nil,
+				EmailTemplates:   mailer.Templates{},
+				EmailBackend:     nil,
+				WebURL:           "http://mock.url",
+			},
+			expectedErr: ErrEmptyEmailBackend,
+		},
+		{
+			app: App{
+				DB:               &gorm.DB{},
+				Clock:            clock.NewMock(),
+				StripeAPIBackend: nil,
+				EmailTemplates:   mailer.Templates{},
+				EmailBackend:     &testutils.MockEmailbackendImplementation{},
+				WebURL:           "",
+			},
+			expectedErr: ErrEmptyWebURL,
+		},
+	}
+
+	for idx, tc := range testCases {
+		t.Run(fmt.Sprintf("test case %d", idx), func(t *testing.T) {
+			_, err := NewRouter(&tc.app)
+
+			assert.Equal(t, errors.Cause(err), tc.expectedErr, "error mismatch")
 		})
 	}
 }