+++ pkg/server/mailer/mailer.go
@@ -22,43 +22,100 @@ package mailer
 import (
 	"bytes"
 	"fmt"
-	"html/template"
-	"os"
-	"path"
-	"strconv"
+	htemplate "html/template"
+	"io"
+	ttemplate "text/template"
 
 	"github.com/aymerick/douceur/inliner"
 	"github.com/gobuffalo/packr/v2"
 	"github.com/pkg/errors"
-	"gopkg.in/gomail.v2"
 )
 
-// Email represents email to be sent out
-type Email struct {
-	from    string
-	to      []string
-	subject string
-	Body    string
-}
-
 var (
-	// T is a map of templates
-	T = map[string]*template.Template{}
 	// EmailTypeResetPassword represents a reset password email
 	EmailTypeResetPassword = "reset_password"
 	// EmailTypeWeeklyDigest represents a weekly digest email
 	EmailTypeWeeklyDigest = "digest"
 	// EmailTypeEmailVerification represents an email verification email
-	EmailTypeEmailVerification = "email_verification"
+	EmailTypeEmailVerification = "verify_email"
+	// EmailTypeWelcome represents an welcome email
+	EmailTypeWelcome = "welcome"
 )
 
-func getTemplatePath(templateDirPath, filename string) string {
-	return path.Join(templateDirPath, fmt.Sprintf("%s.html", filename))
+var (
+	// EmailKindHTML is the type of html email
+	EmailKindHTML = "html"
+	// EmailKindHTML is the type of text email
+	EmailKindText = "text"
+)
+
+// template is the common interface shared between Template from
+// html/template and text/template
+type template interface {
+	Execute(wr io.Writer, data interface{}) error
 }
 
-// initTemplate returns a template instance by parsing the template with the
+// Templates holds the parsed email templates
+type Templates map[string]template
+
+func getTemplateKey(name, kind string) string {
+	return fmt.Sprintf("%s.%s", name, kind)
+}
+
+func (tmpl Templates) get(name, kind string) (template, error) {
+	key := getTemplateKey(name, kind)
+	t := tmpl[key]
+	if t == nil {
+		return nil, errors.Errorf("unsupported template '%s' with type '%s'", name, kind)
+	}
+
+	return t, nil
+}
+
+func (tmpl Templates) set(name, kind string, t template) {
+	key := getTemplateKey(name, kind)
+	tmpl[key] = t
+}
+
+// NewTemplates initializes templates
+func NewTemplates(srcDir *string) Templates {
+	var box *packr.Box
+
+	if srcDir != nil {
+		box = packr.Folder(*srcDir)
+	} else {
+		box = packr.New("emailTemplates", "./templates/src")
+	}
+
+	weeklyDigestHTML, err := initHTMLTmpl(box, EmailTypeWeeklyDigest)
+	if err != nil {
+		panic(errors.Wrap(err, "initializing weekly digest template"))
+	}
+	welcomeText, err := initTextTmpl(box, EmailTypeWelcome)
+	if err != nil {
+		panic(errors.Wrap(err, "initializing welcome template"))
+	}
+	verifyEmailText, err := initTextTmpl(box, EmailTypeEmailVerification)
+	if err != nil {
+		panic(errors.Wrap(err, "initializing email verification template"))
+	}
+	passwordResetText, err := initTextTmpl(box, EmailTypeResetPassword)
+	if err != nil {
+		panic(errors.Wrap(err, "initializing password reset template"))
+	}
+
+	T := Templates{}
+	T.set(EmailTypeWeeklyDigest, EmailKindHTML, weeklyDigestHTML)
+	T.set(EmailTypeResetPassword, EmailKindText, passwordResetText)
+	T.set(EmailTypeEmailVerification, EmailKindText, verifyEmailText)
+	T.set(EmailTypeWelcome, EmailKindText, welcomeText)
+
+	return T
+}
+
+// initHTMLTmpl returns a template instance by parsing the template with the
 // given name along with partials
-func initTemplate(box *packr.Box, templateName string) (*template.Template, error) {
+func initHTMLTmpl(box *packr.Box, templateName string) (template, error) {
 	filename := fmt.Sprintf("%s.html", templateName)
 
 	content, err := box.FindString(filename)
@@ -74,7 +131,7 @@ func initTemplate(box *packr.Box, templateName string) (*template.Template, erro
 		return nil, errors.Wrap(err, "reading footer template")
 	}
 
-	t := template.New(templateName)
+	t := htemplate.New(templateName)
 	if _, err = t.Parse(content); err != nil {
 		return nil, errors.Wrap(err, "parsing template")
 	}
@@ -88,114 +145,44 @@ func initTemplate(box *packr.Box, templateName string) (*template.Template, erro
 	return t, nil
 }
 
-// InitTemplates initializes templates
-func InitTemplates(srcDir *string) {
-	var box *packr.Box
+// initTextTmpl returns a template instance by parsing the template with the given name
+func initTextTmpl(box *packr.Box, templateName string) (template, error) {
+	filename := fmt.Sprintf("%s.txt", templateName)
 
-	if srcDir != nil {
-		box = packr.Folder(*srcDir)
-	} else {
-		box = packr.New("emailTemplates", "./templates/src")
-	}
-
-	weeklyDigestTmpl, err := initTemplate(box, EmailTypeWeeklyDigest)
+	content, err := box.FindString(filename)
 	if err != nil {
-		panic(errors.Wrap(err, "initializing weekly digest template"))
-	}
-	emailVerificationTmpl, err := initTemplate(box, EmailTypeEmailVerification)
-	if err != nil {
-		panic(errors.Wrap(err, "initializing email verification template"))
-	}
-	passwowrdResetTmpl, err := initTemplate(box, EmailTypeResetPassword)
-	if err != nil {
-		panic(errors.Wrap(err, "initializing password reset template"))
+		return nil, errors.Wrap(err, "reading template")
 	}
 
-	T[EmailTypeWeeklyDigest] = weeklyDigestTmpl
-	T[EmailTypeEmailVerification] = emailVerificationTmpl
-	T[EmailTypeResetPassword] = passwowrdResetTmpl
+	t := ttemplate.New(templateName)
+	if _, err = t.Parse(content); err != nil {
+		return nil, errors.Wrap(err, "parsing template")
+	}
+
+	return t, nil
 }
 
-// NewEmail returns a pointer to an Email struct with the given data
-func NewEmail(from string, to []string, subject string) *Email {
-	return &Email{
-		from:    from,
-		to:      to,
-		subject: subject,
-	}
-}
-
-type dialerParams struct {
-	Host     string
-	Port     int
-	Username string
-	Password string
-}
-
-func getSMTPParams() (*dialerParams, error) {
-	portStr := os.Getenv("SmtpPort")
-	port, err := strconv.Atoi(portStr)
+// Execute executes the template with the given name with the givn data
+func (tmpl Templates) Execute(name, kind string, data interface{}) (string, error) {
+	t, err := tmpl.get(name, kind)
 	if err != nil {
-		return nil, errors.Wrap(err, "parsing SMTP port")
-	}
-
-	p := &dialerParams{
-		Host:     os.Getenv("SmtpHost"),
-		Port:     port,
-		Username: os.Getenv("SmtpUsername"),
-		Password: os.Getenv("SmtpPassword"),
-	}
-
-	return p, nil
-}
-
-// Send sends the email
-func (e *Email) Send() error {
-	// If not production, never actually send an email
-	if os.Getenv("GO_ENV") != "PRODUCTION" {
-		fmt.Println("Not sending email because not production")
-		fmt.Println(e.subject, e.to, e.from)
-		// fmt.Println("Body", e.Body)
-		return nil
-	}
-
-	m := gomail.NewMessage()
-	m.SetHeader("From", e.from)
-	m.SetHeader("To", e.to...)
-	m.SetHeader("Subject", e.subject)
-	m.SetBody("text/html", e.Body)
-
-	p, err := getSMTPParams()
-	if err != nil {
-		return errors.Wrap(err, "getting dialer params")
-	}
-
-	d := gomail.NewPlainDialer(p.Host, p.Port, p.Username, p.Password)
-	if err := d.DialAndSend(m); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// ParseTemplate sets the email body by parsing the file at the given path,
-// evaluating all partials and inlining CSS rules
-func (e *Email) ParseTemplate(templateName string, data interface{}) error {
-	t := T[templateName]
-	if t == nil {
-		return errors.Errorf("unsupported template '%s'", templateName)
+		return "", errors.Wrap(err, "getting template")
 	}
 
 	buf := new(bytes.Buffer)
 	if err := t.Execute(buf, data); err != nil {
-		return errors.Wrap(err, "executing the template")
+		return "", errors.Wrap(err, "executing the template")
 	}
 
-	html, err := inliner.Inline(buf.String())
-	if err != nil {
-		return errors.Wrap(err, "inlining the css rules")
+	// If HTML email, inline the CSS rules
+	if kind == EmailKindHTML {
+		html, err := inliner.Inline(buf.String())
+		if err != nil {
+			return "", errors.Wrap(err, "inlining the css rules")
+		}
+
+		return html, nil
 	}
 
-	e.Body = html
-	return nil
+	return buf.String(), nil
 }